+++ src/main/java/babysnoozer/events/SetSnoozingStartPosEvent.java
@@ -1,13 +1,13 @@
 package babysnoozer.events;
 
 public class SetSnoozingStartPosEvent {
-    private final long startPos;
+    private final int startPos;
 
-    public SetSnoozingStartPosEvent(long startPos) {
+    public SetSnoozingStartPosEvent(int startPos) {
         this.startPos = startPos;
     }
 
-    public long getStartPos() {
+    public int getStartPos() {
         return startPos;
     }
 }
